feat: subtract image first copy the image
BREAKING CHANGES:
When subtracting an image the original image is first copied.
Therefore when selecting 'channels' the unselected channels are copied and the difference is only calculated for the selected channels.
The option bitDepth has been removed for subtractImage. While this option makes sense when adding images it seems useless when subtracting them.
Closes https://github.com/image-js/image-js/issues/542