feat!: roiLayer improve speed of mapIDs calculation
The use of an object to calculate the mapIDs is relatively slow.  The algorithm was changed and should not have any impact on functionnalities if no hack were used to create a roiMap (by default the data of an roiMap is Int16Array and should be like that). If there is a large number (over 1000 rois) this change increase the algorithm of a factor 10 (from 1s to 0.1s).